requirements_dev.txt
-r requirements.txt
mockredispy==2.9.0.10
doublex==1.8.2
mock==1.0.1